routes fix
moved news routes to its own file
