+++ src/routes/auth.js
@@ -1,0 +1,11 @@
+const express = require("express");
+const router = express.Router();
+const authController = require('../controllers/auth-controller')
+
+
+
+router.post("/login", authController.login)
+
+
+
+module.exports = router
