--- src/routes/entries.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const entriesController = require('../controllers/entries-controller');
-
-router.get("/news", entriesController.getNewsEntries);
-
-module.exports = router;
