+++ src/routes/index.js
@@ -1,12 +1,16 @@
 const express = require("express");
 const router = express.Router();
 const usersRoutes = require("./users");
+const authRoutes =require("./auth")
 const organizationsRoutes = require('./organization')
-const entriesRoutes = require('./entries');
+const newsRoutes = require('./news');
 
-router.use("/auth", usersRoutes);
+router.use("/users", usersRoutes)
+
+router.use("/auth", authRoutes );
 
 router.use("/organizations", organizationsRoutes)
 
-router.use("/", entriesRoutes);
+router.use("/news", newsRoutes);
+
 module.exports = router;
