+++ src/routes/users.js
@@ -2,8 +2,10 @@ const express = require("express");
 const router = express.Router();
 const userController = require('../controllers/users-controllers')
 
-router.get("/users");
-router.post("/login", userController.login)
+router.get("/");
+
+
+
 
 
 module.exports = router;
